refactor(openai): isolate embedding mapping
Move the existing embedding request and response path out of the compatible Chat transport file. This creates a provider-owned operation boundary without changing the accepted input or wire behavior.
